Refactor remaining projects
Includes renaming project-names, namespace, and references.
Tried to do this one project at a time, but did not like leaving the project in a non-compile state.
