+++ src/DnxFlash/InMemoryMessageProvider.cs
@@ -1,4 +1,4 @@
-﻿namespace MarqueeMessenger
+﻿namespace DnxFlash
 {
     public class InMemoryMessageProvider : IMessageProvider
     {
