+++ test/DnxFlash.Test/InMemoryMessageProviderTest.cs
@@ -1,6 +1,6 @@
 ﻿using Xunit;
 
-namespace MarqueeMessenger.Tests
+namespace DnxFlash.Test
 {
     public class InMemoryMessageProviderTest
     {
